Amélioration du site (#39)
* Add reset password

* Add delete account

* Fixe temporairement le converage demandé

* Fixe temporairement le converage demandé 2

* Ajout de la vérification du mot de passe

* Fix #36 -> suppression des données persistantes + Fix simple search qui ne fonctionnait pas

* Add stats v1

* Ajout de la fonctionnalité de suppression de compte avec vérification d'utilisateur, mise à jour des statistiques de jeu et amélioration de l'interface utilisateur.

* Mise à jour du formulaire des joueurs : ajout du joueur connecté en premier avec un champ en lecture seule pour son nom d'utilisateur.

* Optimisation de la récupération des données de jeu : utilisation de TruncDate pour les dates de jeu distinctes dans les statistiques.

* Mise à jour du README pour marquer les tâches terminées et ajout de nouvelles tâches à réaliser. Amélioration de la méthode de calcul de la note dans le modèle des jeux. Ajustement de l'affichage des tags dans la recherche avancée pour montrer un nombre réduit de tags. Tri des tags par priorité dans la vue de recherche avancée.

* Ajout de l'affichage des extensions de jeu dans la page de détails du jeu et mise à jour de la vue pour inclure les extensions dans le contexte.

* Ajout des stats !!

* Fix ajout historique

* Fix bug de doublons

* Fix bugs

* message d'erreur en rouge

* Supprimer la vérification redondante de l'utilisateur lors de l'ajout de joueurs

* Add icon

---------

+++ website/authentification/views.py
@@ -7,6 +7,7 @@ from django.contrib.auth.decorators import login_required
 
 from authentification import forms
 from django.forms import ValidationError
+from authentification.models import User
 
 def signup(request):
     form = forms.SignupForm()
@@ -80,10 +81,13 @@ def reset_password(request):
     return render(request, 'authentification/reset_password.html', context={'form': form, 'message': message})
 
 @login_required
-def delete_account(request):
+def delete_account(request, id):
+    user = User.objects.get(id=id)
+    if request.user != user and not request.user.is_superuser:
+        return redirect('home')
     if request.method == 'POST':
-        user = request.user
-        logout(request)
+        if request.user == user:
+            logout(request)
         user.delete()
         return redirect('home')
     return render(request, 'authentification/delete_account.html')
